+++ entroppy/__init__.py
@@ -8,7 +8,7 @@ from entroppy.processing import run_pipeline
 from entroppy.utils.debug import DebugTypoMatcher
 from entroppy.utils.logging import setup_logger
 
-__version__ = "0.5.1"
+__version__ = "0.5.3"
 __all__ = [
     "BoundaryType",
     "Config",